+++ client/app/src/main/java/com/example/witt/data/source/remote/signup/SignUpDataSource.kt
@@ -1,7 +1,7 @@
 package com.example.witt.data.source.remote.signup
 
-import com.example.witt.data.model.auth.request.SignUpRequest
-import com.example.witt.data.model.auth.response.SignUpResponse
+import com.example.witt.data.model.remote.signup.request.SignUpRequest
+import com.example.witt.data.model.remote.signup.response.SignUpResponse
 
 interface SignUpDataSource {
     suspend fun signUp(signUpRequest: SignUpRequest) : Result<SignUpResponse>
