+++ client/app/src/main/java/com/example/witt/data/source/remote/signup/SignUpDataSourceImpl.kt
@@ -1,8 +1,8 @@
 package com.example.witt.data.source.remote.signup
 
 import com.example.witt.data.api.SignUpService
-import com.example.witt.data.model.auth.request.SignUpRequest
-import com.example.witt.data.model.auth.response.SignUpResponse
+import com.example.witt.data.model.remote.signup.request.SignUpRequest
+import com.example.witt.data.model.remote.signup.response.SignUpResponse
 import com.example.witt.di.IoDispatcher
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
