--- client/app/src/main/java/com/example/witt/domain/model/auth/SignUpModel.kt
@@ -1,8 +1,0 @@
-package com.example.witt.domain.model.auth
-
-import java.security.cert.CertPathValidatorException
-
-data class SignUpModel (
-    val status: Boolean,
-    val reason: String
-        )
