+++ client/app/src/main/java/com/example/witt/domain/model/remote/signup/DuplicateEmailModel.kt
@@ -1,4 +1,4 @@
-package com.example.witt.domain.model.user
+package com.example.witt.domain.model.remote.signup
 
 data class DuplicateEmailModel (
     val status : Boolean,
