+++ client/app/src/main/java/com/example/witt/domain/model/remote/signup/SignUpModel.kt
@@ -1,0 +1,6 @@
+package com.example.witt.domain.model.remote.signup
+
+data class SignUpModel (
+    val status: Boolean,
+    val reason: String
+        )
