+++ client/app/src/main/java/com/example/witt/domain/use_case/remote/DuplicateEmail.kt
@@ -1,9 +1,7 @@
 package com.example.witt.domain.use_case.remote
 
-import com.example.witt.domain.model.user.DuplicateEmailModel
+import com.example.witt.domain.model.remote.signup.DuplicateEmailModel
 import com.example.witt.domain.repository.UserRepository
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
 class DuplicateEmail @Inject constructor(
